+++ example/lib/app/splash_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import '../common/animated_progress_indicator.dart';
 
 class SplashScreen extends StatelessWidget {
-  final double progress;
+  final double? progress;
 
   const SplashScreen({super.key, required this.progress});
 
@@ -22,9 +22,7 @@ class SplashScreen extends StatelessWidget {
             width: 100,
             child: AnimatedProgressIndicator(
               value: progress,
-              builder:
-                  (value) =>
-                      LinearProgressIndicator(value: value == 0 ? null : value),
+              builder: (value) => LinearProgressIndicator(value: value),
             ),
           ),
         ),