+++ lib/src/scope/scope.dart
@@ -4,13 +4,17 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import 'scope_deps.dart';
-import 'scope_deps_state.dart';
 
 part 'scope_config.dart';
 part 'scope_content.dart';
+part 'scope_deps_state.dart';
 part 'scope_log.dart';
 
-typedef ScopeOnInitCallback<P extends Object?> = Widget Function(P progress);
+typedef ScopeInitFunction<P extends Object, D extends ScopeDeps>
+    = Stream<ScopeInitState<P, D>> Function(BuildContext);
+
+typedef ScopeOnInitCallback<P extends Object> = Widget Function(P? progress);
+
 typedef ScopeOnErrorCallback = Widget Function(
     Object error, StackTrace stackTrace);
 
@@ -18,116 +22,81 @@ typedef ScopeOnErrorCallback = Widget Function(
 ///
 /// Used to create a subtree in the widget tree that guarantees that all
 /// necessary initialized dependencies for widgets will exist.
-///
-/// The final result may look something like this:
-///
-/// ```dart
-/// @override
-/// Widget build(BuildContext context) {
-///   return AppScope(
-///     init: AppScopeDeps.init,
-///     initialization: (step) => SplashScreen(step),
-///     initialized: () => HomeScreen(),
-///   );
-/// }
-///
-/// final class AppScopeDeps implements ScopeDeps {
-///   final SomeController someController;
-///
-///   AppScopeDeps({
-///     required this.someController,
-///   });
-///
-///   static Stream<ScopeProgress<AppScopeDeps>> init(
-///     BuildContext context,
-///   ) async* {
-///     SomeController? someController;
-///     var isInitialized = false;
-///
-///     try {
-///       yield ScopeInit('create $SomeController');
-///       someController = SomeController();
-///       await someController.init();
-///
-///       yield ScopeReady(
-///         AppScopeDeps(
-///           someController: someController,
-///         ),
-///       );
-///
-///       isInitialized = true;
-///     } finally {
-///       if (!isInitialized) {
-///         someController?.dispose();
-///       }
-///     }
-///   }
-///
-///   @override
-///   Future<void> dispose() async {
-///     await someController.dispose;
-///   }
-/// }
-///
-/// class HomeScreen extends StatelessWidget {
-///   @override
-///   Widget build(BuildContext context) {
-///     final someController = AppScope.of(context).someController;
-///     ...
-///   }
-/// }
-/// ```
-abstract base class Scope<
-    S extends Scope<S, P, D, C>,
-    P extends Object?,
-    D extends ScopeDeps,
-    C extends ScopeContent<S, P, D, C>> extends StatefulWidget {
-  final P initialStep;
-  final Stream<ScopeInitState<P, D>> Function(BuildContext context) init;
+abstract base class Scope<S extends Scope<S, D, C>, D extends ScopeDeps,
+    C extends ScopeContent<S, D, C>> extends StatefulWidget {
+  final ScopeInitFunction<Object, D> _init;
 
-  const Scope({super.key, required this.initialStep, required this.init});
-
-  static C of<
-          S extends Scope<S, P, D, C>,
-          P extends Object?,
-          D extends ScopeDeps,
-          C extends ScopeContent<S, P, D, C>>(BuildContext context) =>
-      maybeOf<S, P, D, C>(context) ??
-      (throw Exception('$C not found in the context'));
-
-  static C? maybeOf<
-          S extends Scope<S, P, D, C>,
-          P extends Object?,
-          D extends ScopeDeps,
-          C extends ScopeContent<S, P, D, C>>(BuildContext context) =>
-      _Scope.maybeOf<S, P, D, C>(
-        context,
-        listen: false,
-      )?._contentKey.currentState;
+  const Scope({super.key, required final ScopeInitFunction<Object, D> init})
+      : _init = init;
 
   /// Quick access to parameters passed in scope.
   ///
-  /// Returns scope widget.
+  /// Returns [Scope].
   //
   // Insert the following code into your scope for more convenient use:
   //
   // ```dart
-  // static $YourScope of(BuildContext context, {bool listen = true}) =>
-  //     Scope.paramsOf<$YourScope, $YourScopeDeps>(context, listen: listen);
+  // static $YourScope paramsOf(BuildContext context, {bool listen = true}) =>
+  //     Scope.paramsOf<$YourScope, $YourScopeDeps, $YourScopeContent>(context, listen: listen);
   // ```
-  static S paramsOf<
-              S extends Scope<S, P, D, C>,
-              P extends Object?,
-              D extends ScopeDeps,
-              C extends ScopeContent<S, P, D, C>>(BuildContext context,
+  static S paramsOf<S extends Scope<S, D, C>, D extends ScopeDeps,
+              C extends ScopeContent<S, D, C>>(BuildContext context,
           {bool listen = true}) =>
-      _Scope.maybeOf<S, P, D, C>(context, listen: listen)?.widget ??
+      _Scope.maybeOf<S, D, C>(context, listen: listen)?.widget ??
       (throw Exception('$S not found in the context'));
 
   bool updateParamsShouldNotify(S oldWidget);
 
-  Widget onInit(P progress);
+  /// Quick access to the scope.
+  ///
+  /// Returns [ScopeContent]. Dependencies are in [ScopeContent.deps].
+  //
+  // Insert the following code into your scope for more convenient use:
+  //
+  // ```dart
+  // static $YourScopeContent of(BuildContext context) =>
+  //     Scope.of<$YourScope, $YourScopeDeps, $YourScopeContent>(context);
+  // ```
+  static C of<S extends Scope<S, D, C>, D extends ScopeDeps,
+          C extends ScopeContent<S, D, C>>(BuildContext context) =>
+      maybeOf<S, D, C>(context) ??
+      (throw Exception('$C not found in the context'));
 
+  static C? maybeOf<S extends Scope<S, D, C>, D extends ScopeDeps,
+          C extends ScopeContent<S, D, C>>(BuildContext context) =>
+      _Scope.maybeOf<S, D, C>(
+        context,
+        listen: false,
+      )?._contentKey.currentState;
+
+  /// Метод для построения виджета прогресса инициализации.
+  ///
+  /// Чтобы не вносить тип прогресса в определение [Scope] и, соответственно,
+  /// всех остальных виджетов, зависящих от него, тип прогресса сделан
+  /// универсальным: [Object]. Т.е. может быть любым типом, кроме `null`.
+  ///
+  /// [!WARNING] `null` оставлен для первого запуска [Scope.onInit]
+  /// и обязательно должен быть отработан как этап до начала инициализации.
+  ///
+  /// В функции инициализации тип указывается в описании класса
+  /// [ScopeInitState]:
+  ///
+  /// ```dart
+  /// Stream<ScopeInitState<double, MyDeps>> init(BuildContext context) {...}
+  /// ```
+  ///
+  /// Но метод [onInit] не может автоматически принять этот тип. Необходимо
+  /// самостоятельно конвертировать значение в нужный тип, не забывая про
+  /// первый `null`:
+  ///
+  /// ```dart
+  /// Widget onInit(Object? progress) =>
+  ///     LinearProgressIndicator(value: value as double?)
+  ///
+  /// ```
+  Widget onInit(Object? progress);
+
+  /// Метод для построения виджета ошибки инициализации.
   Widget onError(Object error, StackTrace stackTrace);
 
   C createContent();
@@ -136,45 +105,29 @@ abstract base class Scope<
 
   @override
   @nonVirtual
-  State<S> createState() => _ScopeState<S, P, D, C>();
+  State<S> createState() => _ScopeState<S, D, C>();
 
   @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-
-    properties.add(
-      MessageProperty(
-        'initialStep',
-        '$initialStep',
-        style: DiagnosticsTreeStyle.singleLine,
-        level: DiagnosticLevel.info,
-      ),
-    );
-  }
-
-  @override
-  String toStringShort() => objectRuntimeType(this, '${Scope<S, P, D, C>}');
+  String toStringShort() => objectRuntimeType(this, '${Scope<S, D, C>}');
 }
 
-base class _ScopeState<S extends Scope<S, P, D, C>, P extends Object?,
-    D extends ScopeDeps, C extends ScopeContent<S, P, D, C>> extends State<S> {
+base class _ScopeState<S extends Scope<S, D, C>, D extends ScopeDeps,
+    C extends ScopeContent<S, D, C>> extends State<S> {
   final _contentKey = GlobalKey<C>();
   StreamSubscription<void>? _subscription;
-  late ScopeDepsState<P, D> _state;
+  _ScopeDepsState<Object, D> _state = _ScopeInitial();
 
   @override
   void initState() {
     super.initState();
 
-    _state = ScopeProgress(widget.initialStep);
-
-    _subscription = widget.init(context).listen(
+    _subscription = widget._init(context).listen(
       (state) {
         switch (state) {
-          case ScopeProgress<P, D>(:final value):
+          case ScopeProgress(:final value):
             _debug('$S.init', value);
 
-          case ScopeReady<P, D>():
+          case ScopeReady():
             _unsubscribe();
             _debug('$S.init', '$D is ready');
         }
@@ -194,7 +147,7 @@ base class _ScopeState<S extends Scope<S, P, D, C>, P extends Object?,
         );
 
         setState(() {
-          _state = ScopeError(error, stackTrace);
+          _state = _ScopeError(error, stackTrace);
         });
       },
       onDone: () {
@@ -224,17 +177,19 @@ base class _ScopeState<S extends Scope<S, P, D, C>, P extends Object?,
 
   @override
   Widget build(BuildContext _) {
-    return _Scope<S, P, D, C>(
+    return _Scope<S, D, C>(
       scopeState: this,
       child: switch (_state) {
-        ScopeProgress(:final value) => widget.onInit(value),
-        ScopeError(:final error, :final stackTrace) => widget.onError(
+        _ScopeInitial(:final value) ||
+        ScopeProgress(:final Object? value) =>
+          widget.onInit(value),
+        _ScopeError(:final error, :final stackTrace) => widget.onError(
             error,
             stackTrace,
           ),
         ScopeReady(:final deps) => widget.wrapContent(
             deps,
-            _ScopeContent<S, P, D, C>(
+            _ScopeContent<S, D, C>(
               key: _contentKey,
               deps: deps,
               createContent: widget.createContent,
@@ -245,38 +200,34 @@ base class _ScopeState<S extends Scope<S, P, D, C>, P extends Object?,
   }
 
   @override
-  String toStringShort() => '${_ScopeState<S, P, D, C>}(_state: $_state)';
+  String toStringShort() => '${_ScopeState<S, D, C>}(_state: $_state)';
 }
 
-class _Scope<
-    S extends Scope<S, P, D, C>,
-    P extends Object?,
-    D extends ScopeDeps,
-    C extends ScopeContent<S, P, D, C>> extends InheritedWidget {
-  final _ScopeState<S, P, D, C> scopeState;
+class _Scope<S extends Scope<S, D, C>, D extends ScopeDeps,
+    C extends ScopeContent<S, D, C>> extends InheritedWidget {
+  final _ScopeState<S, D, C> scopeState;
   final S scope;
 
   _Scope({super.key, required this.scopeState, required super.child})
       : scope = scopeState.widget;
 
-  static _ScopeState<S, P, D, C>? maybeOf<
-              S extends Scope<S, P, D, C>,
-              P extends Object?,
+  static _ScopeState<S, D, C>? maybeOf<
+              S extends Scope<S, D, C>,
               D extends ScopeDeps,
-              C extends ScopeContent<S, P, D, C>>(BuildContext context,
+              C extends ScopeContent<S, D, C>>(BuildContext context,
           {required bool listen}) =>
       listen
           ? context
-              .dependOnInheritedWidgetOfExactType<_Scope<S, P, D, C>>()
+              .dependOnInheritedWidgetOfExactType<_Scope<S, D, C>>()
               ?.scopeState
           : context
-              .getInheritedWidgetOfExactType<_Scope<S, P, D, C>>()
+              .getInheritedWidgetOfExactType<_Scope<S, D, C>>()
               ?.scopeState;
 
   @override
-  bool updateShouldNotify(_Scope<S, P, D, C> oldWidget) =>
+  bool updateShouldNotify(_Scope<S, D, C> oldWidget) =>
       scope.updateParamsShouldNotify(oldWidget.scope);
 
   @override
-  String toStringShort() => '${_Scope<S, P, D, C>}';
+  String toStringShort() => '${_Scope<S, D, C>}';
 }